Style "select size" button
Styled the select size button so that it is on the top of the grid.
Changed the border, size, and added a hover effect.

Also removed the div container that the grid was in. This was so that I
I could properly change the button placement using flexbox. Edited the
rest of the CSS and JavaScript code accordingly for this change.